requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
flake8==3.9.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
